Spree Windows Client
Spree, our in house automated software promotion has migrated to use a
RestFUL API, this is the windows client to utilize the new API and give
our windows servers the ability to auto promote code deployments using
Spree. Also I split off the logic to actually do the promotion. How I
have it working now is. All the client does is first checkin to the
server, then it will grab a script name from spree and run that script.
This will allow it to do more than just promotions. I can have it look
for and run any script I tell it too. which will enable us in the future
to automate more tasks without having to change the client and have it
evolve into some monolithic beast of a script.